refactor(client): improve rubocop compliance
This commit refactors `client.rb` to improve rubocop compliance. This
will be an ongoing process; this commit does not seek full compliance
with rubocop, nor should it.

The biggest change in this commit is `freeze`-ing `DEFAULTS`. This hash
is meant to be a constant. This change enforces the ALL_CAPS semantics
such that modifying this value will result in a `FrozenError`:

```ruby
irb(main):001:0> H = {a: 1}.freeze
=> {:a=>1}
irb(main):002:0> H[:a] = 2
Traceback (most recent call last):
        4: from /Users/Eric/.rbenv/versions/2.6.3/bin/irb:23:in `<main>'
        3: from /Users/Eric/.rbenv/versions/2.6.3/bin/irb:23:in `load'
        2: from
/Users/Eric/.rbenv/versions/2.6.3/lib/ruby/gems/2.6.0/gems/irb-1.0.0/exe/irb:11:in
`<top (required)>'
        1: from (irb):2
FrozenError (can't modify frozen Hash)
```

Double quoted strings have been exchanged for single quoted strings
(except for where interpolation is used).

Error messages have been refactored to:
- reduce their line length
- prepare for unified error messaging

Why? Consider these assignments to be placeholders. This way we can
talk about whether something should be called a `host_error` or a
`domain_error` and get a feel for it while preparing to unify error
messaging under an `ImgixError` or `Error` module. Once we have
settled on the details, the new error values can be swapped out
without ambiguity. Unifying error messaging this way will pull
most of the strings out of this file.

**Ambiguity**:
- line 27, `ixlib` cannot be found
- line 52, `path` is never used
  - it is used and passed an argument in tests, though it doesn't get
used
  - this will be refactored out in the next pass
- lines 32 thru 50 will need another pass and a dedicated PR
- line 34, there was a double (perhaps triple negation, i.e. `unless
!!(@api_key`))
  - double check that I've refactored in favor of our semantics

- [x] all tests pass