refactor(imgix.rb): freeze immutable objects
This commit freezes two constants defines in `imgix.rb`. Named
constants are expected to be immutable, any mutation to these
structures should be an error.

The naming of `TARGET_WIDTHS` suggests that it is a constant. However,
this is not the case. It is a lambda, when called with different
arguments produces different results based on those different args. It
is not constant.

Further refactoring should be reserved until after some discussion on
what to do here. Initially, I am a fan of converting `TARGET_WIDTHS`
into an array while providing a new function named for generating
target widths.

**Proposed:**

```ruby
TARGET_WIDTHS = [100, ..., 8192].freeze

def generate_target_widths()
    # -- snip --
end
```

The above change will not be included because it requires changing
more than one file.

Holding off on further changes.