perf: reduce gem size
The purpose of this commit is to reduce the size of the gem users
install. Prior to this commit, _everything_ in the repo will be
installed upon `gem` or `bundle install`. Now, only pertinent files
are installed.

These files are listed below:
```text
.
├── CHANGELOG.md
├── Contributing.markdown
├── Gemfile
├── Gemfile.lock
├── LICENSE
├── README.md
├── Rakefile
├── imgix.gemspec
└── lib
    ├── imgix
    │   ├── client.rb
    │   ├── param_helpers.rb
    │   ├── path.rb
    │   └── version.rb
    └── imgix.rb
```

The size diff is:

```text
% du -sh imgix-3.4.0/

  148K    imgix-3.4.0

% du -sh imgix-3.4.0

  80K    imgix-3.4.0
```

By not including tests we can save users 68K bytes.