feat: Allow domain to have a port
This allows us to use imgix lib without much changes on local
environment. We can then have the source be something like
"localhost.me:3000" and it just works, pointing to the right asset but
without any transformations.

Without this, devs who can't use S3 on the local environment have to
resort to hacks on their code to workaround this small regex test, which
is not ideal.